+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/models/events/AbstractDynamicLibEventDisconnection.java
@@ -9,6 +9,8 @@ package com.powsybl.dynawaltz.models.events;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.dynawaltz.DynaWaltzContext;
+import com.powsybl.dynawaltz.parameters.ParameterType;
+import com.powsybl.dynawaltz.xml.ParametersXml;
 import com.powsybl.iidm.network.Identifiable;
 
 import javax.xml.stream.XMLStreamException;
@@ -19,12 +21,16 @@ import javax.xml.stream.XMLStreamWriter;
  */
 public abstract class AbstractDynamicLibEventDisconnection extends AbstractEventModel {
 
+    private static final String EVENT_PREFIX = "Disconnect_";
     private static final String DYNAMIC_MODEL_LIB = "EventSetPointBoolean";
     private static final String DEFAULT_MODEL_LIB = "EventConnectedStatus";
     protected static final String DISCONNECTION_VAR_CONNECT = "event_state1";
 
-    protected AbstractDynamicLibEventDisconnection(Identifiable<?> equipment, double startTime) {
-        super(equipment, startTime);
+    private final boolean disconnect;
+
+    protected AbstractDynamicLibEventDisconnection(Identifiable<?> equipment, double startTime, boolean disconnect) {
+        super(equipment, startTime, EVENT_PREFIX);
+        this.disconnect = disconnect;
     }
 
     @Override
@@ -43,5 +49,11 @@ public abstract class AbstractDynamicLibEventDisconnection extends AbstractEvent
         writer.writeAttribute("lib", context.isWithoutBlackBoxDynamicModel(getEquipment().getId()) ? DEFAULT_MODEL_LIB : DYNAMIC_MODEL_LIB);
         writer.writeAttribute("parFile", getParFile(context));
         writer.writeAttribute("parId", getParameterSetId());
+    }
+
+    @Override
+    protected void writeEventSpecificParameters(XMLStreamWriter writer, DynaWaltzContext context) throws XMLStreamException {
+        ParametersXml.writeParameter(writer, ParameterType.DOUBLE, "event_tEvent", Double.toString(getStartTime()));
+        ParametersXml.writeParameter(writer, ParameterType.BOOL, "event_stateEvent1", Boolean.toString(disconnect));
     }
 }
